feat(rule_based): car decision
